Add uncommitted files from core
