+++ core/src/store/redux/actions.ts
@@ -1,25 +1,7 @@
-import { ReportActionOptions } from "./types";
+import { ReportActionOptions, ReportErrorActionCreatorWithType, DismissErrorActionCreatorWithType } from "./types";
 import typedActionCreatorFactory from "../../typed-action-creator/factory";
-import {Action} from "redux";
 
 const randomIdentifier = () => Math.random().toString(36).substring(2, 15)
-
-export type ReportErrorAction = Action & {
-  error: Error;
-  identifier: string;
-
-  options?: ReportActionOptions;
-}
-
-type ReportErrorActionCreator = (error: Error, options?: ReportActionOptions, identifier?: string) => ReportErrorAction;
-export type ReportErrorActionCreatorWithType = ReportErrorActionCreator & { type: string };
-
-export type DismissErrorAction = Action & {
-  identifier: string;
-}
-
-type DismissErrorActionCreator = (identifier: string) => DismissErrorAction;
-export type DismissErrorActionCreatorWithType = DismissErrorActionCreator & { type: string };
 
 export const reportError : ReportErrorActionCreatorWithType = typedActionCreatorFactory(
   '@Galette/REPORT_ERROR',
