+++ core/src/store/redux/types.ts
@@ -25,3 +25,20 @@ export type ReportActionOptions = {
 export type ErrorModuleState = {
   reportedErrors: ReportedError[];
 }
+
+export type ReportErrorAction = Action & {
+  error: Error;
+  identifier: string;
+
+  options?: ReportActionOptions;
+}
+
+type ReportErrorActionCreator = (error: Error, options?: ReportActionOptions, identifier?: string) => ReportErrorAction;
+export type ReportErrorActionCreatorWithType = ReportErrorActionCreator & { type: string };
+
+export type DismissErrorAction = Action & {
+  identifier: string;
+}
+
+type DismissErrorActionCreator = (identifier: string) => DismissErrorAction;
+export type DismissErrorActionCreatorWithType = DismissErrorActionCreator & { type: string };
